1. clean code; 2. add pixart-control inference code; 3. bug fixed;
