+++ diffusion/data/datasets/__init__.py
@@ -2,5 +2,5 @@ from .SA import SAM
 from .InternalData import InternalData
 from .InternalData_ms import InternalDataMS
 from .Dreambooth import DreamBooth
-from .control_hed import MJHed
+from .pixart_control import InternalDataHed
 from .utils import *
